+++ src/Plugin.php
@@ -27,7 +27,12 @@ use MeCms\Command\Install\RunAllCommand;
 use MeTools\Command\Install\CreateDirectoriesCommand;
 use MeTools\Command\Install\CreateVendorsLinksCommand;
 use MeTools\Command\Install\SetPermissionsCommand;
+use MeTools\Plugin as MeTools;
+use RecaptchaMailhide\Plugin as RecaptchaMailhide;
 use Symfony\Component\Finder\Finder;
+use StopSpam\Plugin as StopSpam;
+use Thumber\Cake\Plugin as Thumber;
+use Tokens\Plugin as Tokens;
 
 /**
  * Plugin class
@@ -52,15 +57,23 @@ class Plugin extends BasePlugin
     public function bootstrap(PluginApplicationInterface $app): void
     {
         $pluginsToLoad = array_merge([
-            'MeTools\Plugin',
-            'RecaptchaMailhide\Plugin',
-            'StopSpam\Plugin',
-            'Thumber\Cake\Plugin',
-            'Tokens\Plugin',
+            MeTools::class,
+            RecaptchaMailhide::class,
+            StopSpam::class,
+            Thumber::class,
+            Tokens::class,
         ], getConfig('default.theme') ? [getConfig('default.theme')] : [], !$this->isCli() ? ['WyriHaximus/MinifyHtml'] : []);
-        foreach ($pluginsToLoad as $plugin) {
+        foreach ($pluginsToLoad as $className) {
             /** @var \Cake\Http\BaseApplication $app */
-            if (!$app->getPlugins()->has($plugin)) {
+            if (!$app->getPlugins()->has($className)) {
+                if ($className == 'WyriHaximus/MinifyHtml') {
+                    $app->addPlugin($className);
+                    continue;
+                }
+
+                $plugin = new $className();
+                $plugin->bootstrap($app);
+                $plugin->disable('bootstrap');
                 $app->addPlugin($plugin);
             }
         }